feat: add Task feature added

+++ lib/widgets/custom_todo_list.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:todo/constants.dart';
 import 'package:todo/models/todo_model.dart';
+
+import '../resources/todo_data.dart';
 
 class TodoList extends StatelessWidget {
   const TodoList({
@@ -7,7 +10,7 @@ class TodoList extends StatelessWidget {
     required this.data,
     required this.todos,
   }) : super(key: key);
-  final data;
+  final TodoData data;
   final List<Todo> todos;
 
   @override
@@ -19,9 +22,12 @@ class TodoList extends StatelessWidget {
                 onPressed: () {
                   data.updateTodo(todos[index]);
                 },
-                icon: Icon(todos[index].isCompleted
-                    ? Icons.circle
-                    : Icons.circle_outlined)),
+                icon: Icon(
+                  todos[index].isCompleted
+                      ? Icons.circle
+                      : Icons.circle_outlined,
+                  color: kBgColor,
+                )),
             title: Text(
               todos[index].task,
               style: TextStyle(
@@ -31,9 +37,31 @@ class TodoList extends StatelessWidget {
             ),
             trailing: IconButton(
                 onPressed: () {
-                  data.deleteTodo(todos[index]);
+                  showDialog<String>(
+                    context: context,
+                    builder: (BuildContext context) => AlertDialog(
+                      title: const Text('Warning'),
+                      content: const Text('Are you sure ?'),
+                      actions: <Widget>[
+                        TextButton(
+                          onPressed: () => Navigator.pop(context),
+                          child: const Text('Cancel'),
+                        ),
+                        TextButton(
+                          onPressed: () async {
+                            data.deleteTodo(todos[index]);
+                            Navigator.pop(context);
+                          },
+                          child: const Text('OK'),
+                        ),
+                      ],
+                    ),
+                  );
                 },
-                icon: const Icon(Icons.cancel_outlined)),
+                icon: const Icon(
+                  Icons.cancel_outlined,
+                  color: Colors.black,
+                )),
           );
         }),
         separatorBuilder: ((context, index) {
